requirements.txt
asgiref==3.10.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-database-url==3.0.1
Django==5.2.7
django-environ==0.12.0
django-extensions==4.1
django-jazzmin==3.0.1
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.11
MarkupSafe==3.0.3
mysqlclient==2.2.7
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pycparser==2.23
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
